Refactor userdata openshift
