+++ api/cmd/userdata-openshift/main.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/golang/glog"
 
+	"github.com/kubermatic/kubermatic/api/pkg/userdata/openshift"
 	"github.com/kubermatic/machine-controller/pkg/userdata/plugin"
 )
 
@@ -13,7 +14,7 @@ func main() {
 	flag.BoolVar(&debug, "debug", false, "Switch for enabling the plugin debugging")
 	flag.Parse()
 
-	var provider = &Provider{}
+	var provider = &openshift.Provider{}
 	var p = plugin.New(provider, debug)
 
 	if err := p.Run(); err != nil {
