+++ api/pkg/userdata/openshift/openshift.go
@@ -1,4 +1,4 @@
-package main
+package openshift
 
 import (
 	"bytes"
